requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
pillow
python-multipart
requests
Pillow>=10.0.0
